Adding 2 to 2 rollover, search and prepopulation
Added the ability to rollover from moodle 2 modules to moodle 2 modules
Auto complete shows moodle 2 modules now. Ajax calls to the sources for this are
completely seperate and therefore if one falls over, you can still rollover using
the other source.
Search functionality added so users can search for specific rollovers
Rollovers on the page when it is loaded are given a best guess. Rather than this
being flat text as a guide, this is now an actual selection of what I hope is the
previous incarnation of the module. This should allow users to just push the
rollover button straight away. When the best guess fails it will put the short
code in as text as before.